% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ClusterCols}
\alias{ClusterCols}
\title{Matching clusters with colours}
\usage{
ClusterCols(x, Data, nrclusters = NULL, cols = NULL, colorComps = NULL)
}
\arguments{
\item{x}{The leaf of a dendrogram.}

\item{Data}{A clustering object.}

\item{nrclusters}{The number of clusters to divide the dendrogram in. Default is NULL.}

\item{cols}{A character vector with the colours to be used. Default is NULL.}

\item{colorComps}{A character vector of a specific set of objects to be coloured. Default is NULL.}
}
\description{
Internal function of \code{ClusterPlot}.
}
