% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Colors1}
\alias{Colors1}
\title{Colour examples}
\format{An object if class \code{"character"}.}
\description{
A vector of HEX codes for the colours used in the examples
}
\examples{
data(Colors1)
}
\keyword{datasets}
