\name{Colors2}
\alias{Colors2}
\docType{data}
\title{
Second example for colors
}
\description{
A data set HEX code for the colors used in the examples.
}
\usage{data("Colors2")}
\format{
  The format is:
 chr [1:8] "#D2691E" "#EE2C2C" "#EEAD0E" "#006400" "#0000EE" ...
}
\keyword{datasets}
