% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f.clustABC.MultiSource}
\alias{f.clustABC.MultiSource}
\title{f.clustABC.MultiSoucre}
\usage{
f.clustABC.MultiSource(res, numclust = NULL, distmeth = 1,
  linkage = "ward", alpha = 0.625, mds = FALSE)
}
\arguments{
\item{res}{Matrix object whose rows are the base clusters determined in each iteration of the ABC algorithm.}

\item{numclust}{The number of clusters. Default is NULL.}

\item{distmeth}{Binary, indicating whether the ABC dissimilarities should be scaled by the total number of simulations. By default distmeth=1 indicating the dissimilarities should be left as is.}

\item{linkage}{Choice of inter group dissimilarity (character) for the final clustering. Defaults to "ward.}

\item{alpha}{The parameter alpha to be used in the "flexible" linkage of the agnes function. Defaults to 0.625 and is only used if the linkage is set to "flexible"}

\item{mds}{Logical, indicating whether and MDS plot of the dissimilarities should be drawn. Default is FALSE}
}
\description{
Internal function of M_ABC: performs the final clustering.
}
