% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f.gsample}
\alias{f.gsample}
\title{f.gsample}
\usage{
f.gsample(zf, ng = 100)
}
\arguments{
\item{zf}{Statistics to be used in the weighting of the rows.}

\item{ng}{The number to resample. Default is 100.}
}
\description{
internal function of M_ABC: samples a predetermined number of rows according to specified weights.
}
