% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f.rmv}
\alias{f.rmv}
\title{f.rmv}
\usage{
f.rmv(x, varonly = FALSE)
}
\arguments{
\item{x}{data variable}

\item{varonly}{Logical. Whether only the variance should be returned (TRUE) or the mean as well (FALSE). Default is FALSE.}
}
\description{
internal function of M_ABC: computes means and variances.
}
