% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f.t}
\alias{f.t}
\title{ff}
\usage{
f.t(x, his = FALSE)
}
\arguments{
\item{x}{data variable}

\item{his}{Logical. Whether a histogram should be plotted. Default is FALSE.}
}
\description{
Internal function of M_ABC: determine weights with two sample t-test.
}
