% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fingerprintMat}
\alias{fingerprintMat}
\title{Fingerprint data}
\format{An object of class \code{"matrix"}.}
\usage{
data(fingerprintMat)
}
\description{
A binary data matrix that contains 250 fingerprints for a set of 56 compounds.
}
\examples{
data(fingerprintMat)
}
\keyword{datasets}
