% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GeneInfo}
\alias{GeneInfo}
\title{Information of the genes
 
Gene info in a data frame}
\format{A data frame with 3 variables: ENTREZID, SYMBOL and GENENAME}
\description{
Information of the genes
 
Gene info in a data frame
}
\examples{
data(GeneInfo)
}
\keyword{datasets}
