% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneMat}
\alias{geneMat}
\title{Gene expression data}
\format{An object of class \code{"matrix"}.}
\usage{
data(geneMat)
}
\description{
Gene expression data for 2434 genes for a set of 56 compounds.
}
\examples{
data(geneMat)
}
\keyword{datasets}
