% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GS}
\alias{GS}
\title{List of GO Annotations}
\format{A data frame with 3 variables: ENTREZID, SYMBOL and GENENAME}
\description{
A list that contains the GO annotations produced by getGeneSets of the \code{MLP} package for the genes in the geneMat data.
}
\examples{
data(GS)
}
\keyword{datasets}
