\name{IntClust-package}
\alias{IntClust-package}
\docType{package}
\title{
Integrated Data Analysis
}
\description{
The package contains several integrative data methods in which information of 
objects from different data sources can be combined. As a single data source
is limited in its point of view, this provides more insight and the opportunity
to investigate how the variables are interconnected. Clustering techniques are
to be applied to the combined information. For now, only agglomerative 
hierarchical clustering is implemented. Further, differential gene expression
and pathway analysis can be conducted on the clusters. Plotting functions
are available to visualize and compare results of the different methods. 
}
\details{
\tabular{ll}{
Package: \tab IntClust\cr
Type: \tab Package\cr
Version: \tab 0.0.18\cr
Date: \tab 2014-10-30\cr
License: \tab GPL-3\cr
}
}
\author{
Marijke Van Moerbeke

Maintainer: Marijke Van Moerbeke <marijkevanmoerbeke@uhasselt.be>
}
\keyword{ package }
