% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntClust.R
\docType{package}
\name{IntClust}
\alias{IntClust}
\alias{IntClust-package}
\title{Integrated Clustering Methods.}
\description{
The IntClust package contains several multi-source clustering methods for the integration of multiple data sets.
}
