% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LabelCols}
\alias{LabelCols}
\title{Colouring labels}
\usage{
LabelCols(x, Sel1, Sel2 = NULL, col1 = NULL, col2 = NULL)
}
\arguments{
\item{x}{The leaf of a dendrogram.}

\item{Sel1}{The selection of objects to be colored. Default is NULL.}

\item{Sel2}{An optional second selection to be colored. Default is NULL.}

\item{col1}{The color for the first selection. Default is NULL.}

\item{col2}{The color for the optional second selection. Default is NULL.}
}
\description{
Internal function of \code{LabelPlot}.
}
