% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{targetMat}
\alias{targetMat}
\title{Target prediction data}
\format{An object of class \code{"matrix"}.}
\usage{
data(targetMat)
}
\description{
A binary data matrix that contains 477 target predictions for a set of 56 compounds.
}
\examples{
data(targetMat)
}
\keyword{datasets}
