\name{Colors1}
\alias{Colors1}
\docType{data}
\title{
First example for colors
}
\description{
A data set HEX code for the colors used in the examples.
}
\usage{data("Colors1")}
\format{
  The format is:
 chr [1:8] "#8EE5EE" "#D2691E" "#EE2C2C" "#EEAD0E" "#006400" ...
}
\keyword{datasets}
