\name{fingerprintMat}
\alias{fingerprintMat}
\docType{data}
\title{
The fingerprint matrix for the MCF7 data
}
\description{
A binary data set that contains 250 fingerprints for the CMAP MCF7 data.
}
\usage{data("fingerprintMat")}
\format{
  The format is:
 logi [1:56, 1:250] FALSE FALSE FALSE FALSE FALSE FALSE ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:56] "metformin" "phenformin" "phenyl biguanide" "estradiol" ...
  ..$ : chr [1:250] "-2147375257" "-2147119955" "-2146474760" "-2145840573" ...
}
\keyword{datasets}
