\name{GeneInfo}
\alias{GeneInfo}
\docType{data}
\title{
The gene info matrix
}
\description{
The data set contains the entrezIdentifiers, symbols and names
of the used genes in the CMAP MCF7 data.
}
\usage{data("GeneInfo")}
\format{
  The format is:
 chr [1:2434, 1:3] "10001" "100129361" "10015" "100188893" ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:2434] "1" "2" "3" "4" ...
  ..$ : chr [1:3] "ENTREZID" "SYMBOL" "GENENAME"
}
\keyword{datasets}
