\name{GeneInfo}
\alias{GeneInfo}
\docType{data}
\title{
The gene info data frame
}
\description{
The data set contains the entrezIdentifiers, symbols and names
of the used genes in the CMAP MCF7 data.
}
\usage{data("GeneInfo")}
\format{
  The format is:
 'data.frame':	2434 obs. of  3 variables:
 $ ENTREZID: Factor w/ 2434 levels "10001","100129361",..: 1 2 3 4 5 6 7 8 9 10 ...
  ..- attr(*, "names")= chr  "1" "2" "3" "4" ...
 $ SYMBOL  : Factor w/ 2434 levels "AARS","ABCA1",..: 1213 1132 1486 2178 1914 1133 2175 891 2003 1134 ...
  ..- attr(*, "names")= chr  "1" "2" "3" "4" ...
 $ GENENAME: Factor w/ 2434 levels "1-acylglycerol-3-phosphate O-acyltransferase 5 (lysophosphatidic acid acyltransferase, epsilon)",..: 1199 925 1586 2153 1895 2374 2067 908 1983 926 ...
  ..- attr(*, "names")= chr  "1" "2" "3" "4" ...
}
\keyword{datasets}
