\name{geneMat}
\alias{geneMat}
\docType{data}
\title{
The gene expression matrix
}
\description{
The gene expression data for 2434 genes for the compounds in the
CMAP MCF7 data.
}
\usage{data("geneMat")}
\format{
  The format is:
 num [1:2434, 1:56] -0.0772 -0.0698 -0.055 -0.0498 -0.0597 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:2434] "MED6" "LOC100129361" "PDCD6IP" "TOMM6" ...
  ..$ : chr [1:56] "metformin" "phenformin" "phenyl biguanide" "estradiol" ...
}
\keyword{datasets}
