\name{targetMat}
\alias{targetMat}
\docType{data}
\title{
The target prediction matrix
}
\description{
A binary data matrix that contains 477 target predictions for the CMAP
MCF7 data.
}
\usage{data("targetMat")}
\format{
  The format is:
 num [1:56, 1:477] 0 0 0 0 0 0 0 0 0 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:56] "metformin" "phenformin" "phenyl biguanide" "estradiol" ...
  ..$ : chr [1:477] "Arachidonate_15.lipoxygenase" "Estradiol_17.beta.dehydrogenase_2" "Estradiol_17.beta.dehydrogenase_1" "Lanosterol_synthase" ...
}
\keyword{datasets}
